needs more docker
This adds a chrome docker container that can be used
to view tirefi.re in development.

instructions for using it can be found in tirefi.re/docker

should probably register the tirefi namespace on docker registry
so that we can add it as a trusted build as `tirefi/re`.

it probably only works on linux, but hey, that's kinda the point.